Switch CLI page demo from email style to terminal
